refactor(sep41): relocate spec test helpers into sep41 package
  Move the XDR spec helpers from internal/services/sep41_spec_helpers_test.go
  into internal/services/sep41/spec_helpers_test.go, namespacing them like
  the rest of the sep41 files and eliminating the duplicate copies that had
  been living in sep41/validator_test.go.

  protocol_validator_test.go (services package) no longer has access to the
  helpers across the test-package boundary, so its single spec round-trip
  case now builds a minimal 2-entry spec inline — it only asserts entry
  count, so the SEP-41 surface was unnecessary there.
